Check for timeout on Tradfri observations
This change will check whether or not the Tradfri observation had a timeout. A
timeout can occur when a request has been fired, but the observation state -
which is also the object state - has not been updated for more than 15 seconds.
When this happens, the observation will throw an error (currently an empty one)
which in turn makes Home Assistant re-add a new observation on that device.

It can happen that the observations will reset once or twice when starting up
Home Assistant. This happens on slow systems (such as a Raspberry Pi) but
should have no real impact on the system, Home Assistant or any observations.

After thinking the best approach was to check for specific status codes, I came
to the conclusion that these codes were not really meaningful and couldn't be
used for troubleshooting the observations. This should be a better approach.